fix: cast PortNumber to int32 for gateway-api compatibility
gateway-api v1.2.1 defines PortNumber as a named type (not an alias),
which prevents direct comparison with int32. This cast ensures
compatibility across gateway-api versions.

+++ pkg/controller/v1alpha2/llmisvc/router_discovery.go
@@ -458,7 +458,7 @@ func listenerMatchesServicePorts(listener *gwapiv1.Listener, svcPorts []corev1.S
 		return true
 	}
 	for _, sp := range svcPorts {
-		if listener.Port == sp.Port {
+		if int32(listener.Port) == sp.Port {
 			return true
 		}
 	}
